CST/Bindings API changes for backend (#1304)
This PR makes some changes to the Rust side of CST and Bindings Graph
APIs:

- Define an opaque `NodeId` to represent CST node IDs, which wrap the
previous `usize` type. So now `TerminalNode::id()`,
`NonterminalNode::id()` and `Node::id()` return `NodeId`. This improves
type safety when handling node IDs. Note: the WIT/JS bindings are
unchanged and the ids there are still `u32`.
- Add `definition_by_node_id()` and `reference_by_node_id()` methods to
the `BindingGraph` to be able to retrieve `Definition`s and `Reference`s
from a node ID without needing a `Cursor` object. These are required to
allow the backend to use the binding graph effectively.
- Register `Definition`s in the binding graph via both their identifier
node ID and their definiens node ID. Eg. for a contract, the
`Definition` can be retrieved by the ID of the `ContractDefinition` node
or the `Identifier` acting as its name.
- `Definition::id()` now returns the ID of the definiens node instead of
the identifier node. This should make it more straightforward in the
backend to work with `Definition` objects from the binding graph and
internal structures such as a type registry.