Generate an AST from a CST in Rust (#1266)
This PR brings an AST similar to what's available in Typescript to Rust
land, intended to be a good starting point for further semantic analysis
in the backend. The AST model is built from the language definition.

Sequence language elements are translated into ref counted `struct`s,
choices into `enum`s, and separated and repeated into `Vec`s. Redundant
terminal nodes (ie. unique terminals such as punctuation and keywords)
are removed, as well as language elements deprecated in or before
version 0.8.0.

There's still room for improvement to make the resulting API more
ergonomic, eg. flattening structures. There is an L1 language that's
built from the AST which is a proof of concept for it.

Along with the definition of the nodes of the AST, we additionally
generate, as Rust traits that can be extended:
- a visitor that can stop at sequence-type AST nodes
- an mutator that produces a copy of the AST by default, but can be
extended to update parts of the tree
- a builder for the first IR in the chain, that instantiates the AST
from the parsed CST
- a transformer to convert an IR tree into the next IR language in the
chain (currently AST -> L1)