Update binding rules to bind to unique definiens nodes (#1326)
- For Yul variables, we were binding to the
`YulVariableDeclarationStatement`, but that's incorrect because they can
enumerate many variables. We now bind to the variable name itself.
- For tuple deconstruction declarations, we now bind to the tuple member
variant. In practical terms this is a minor change, but it's convenient
as the `TupleMember` is typed as an enum when building an AST, and we
don't save the node ID for it.