Small fixes to generated IR node and visitor types (#1327)
- Call enter/leave visitor methods when accepting choice types. These
methods were added to the visitor trait, but were never called from the
accept functions.
- When generating IR node types for precedence expressions with multiple
operator variants, those are collapsed into a single struct and we lose
track of the operator that acts as a discriminant, because it's normally
a unique terminal. Detect this case and don't remove the terminals in
these cases. Eg. for an `AdditiveExpression` we now have the operator, a
`PLUS` or a `MINUS`.