Register resolved imports when building the `CompilationUnit` on Sourcify tests (#1337)
When building the `CompilationUnit` for Sourcify tests we were missing
the registration of the resolved imports. This information is later used
by the binding graph to link the partial graphs from each file. Without
this, resolving symbols in imported files didn't work properly.

This PR also includes two additional improvements:

- the argument to `--shard-count` is validated and can now accept the
value 256 (it wouldn't previously because it was stored in a `u8`)
- when reporting failures, always print the contract name alongside the
version (this makes it easier to iterate on failing contracts using the
`--contract` option)

---------

Co-authored-by: Omar Tawfik <15987992+OmarTawfik@users.noreply.github.com>