Change the IR model builder `build_*` functions to return an `Option`
The builder should be tolerant to parsing errors for us to be able to use it in
the new binder implementation. So, the `build_*` functions will return `None` if
constructing the IR node is not possible (because there's an
`UNRECOGNIZED`/`MISSING` node in the source CST). Other unexpected sequences of
nodes or labels, or incorrect node kinds, will instead panic as they indicate an
error in the logic of the builder. The assumption is that the parser will
produce a well-formed CST. This also means the new binder will not necessarily
work well with a user-constructed CST, as there are no guarantees on the
correctness of the structure in that case.