Update to rust 1.87 (#1356)
Relevant changes besides trivial linting error fixes:

1. A somewhat controversial change replaced `&Option<T>` into
`Option<&T>`, requiring lots of conversions `.as_ref()`: when
destructing by reference a struct with an `Option<T>` field, you get an
`&Option<T>`, so in order to operate with it, we need to convert it
first (and this happens a lot).

2. The update to public-api is required, otherwise it fails to produce
the documentation.

3. I kept the `_` in `_into_ffi` and such functions for a simple reason:
when I removed it, other linting problems started to pop-up: one about
the use of `self` in `from_` functions, and having unused functions
(when they are not called). So the amount of work was going to be the
same renaming or not these functions, so I decided to keep them as they
are now.

4. Removed `once_cell::Lock` in favor of stdlib drop-in replacement
`LazyLock`.