More efficient Sourcify CI workflows (#1348)
This PR greatly improves the execution time for running the Sourcify CI
workflows. It is especially noticeable when running all chains - you can
see an improvement from
[1:10:35](https://github.com/NomicFoundation/slang/actions/runs/14881875285)
to just
[15:22](https://github.com/manastech/slang/actions/runs/15420586581).
With this improved performance it's much more feasible to run Sourcify
tests on small changes because you won't have to wait an hour to get the
full picture. It also means spending less on Github Actions minutes.

The main thing slowing down these tests was that we were recompiling
`solidity_testing_sourcify` _and_ `infra` for every shard under test.
Not only was this work highly redundant, but it turns out that compiling
`infra` is completely unnecessary for these tests. Compiling
`solidity_testing_sourcify` directly works just as well, and is much
faster.

To avoid this I added a new script called `scripts/bin/with-hermit`.
This script does the same hermit configuration done by
`scripts/bin/infra`, which is necessary to build or run any Slang binary
in CI, and then runs whatever command you pass it. This means that we
can avoid `infra` and compile the testing binary directly.

On top of this I rewrote the workflow `sourcify_single_chain` so that it
compiles the testing binary in a separate job before starting
`singleShard`. This means that the binary gets compiled once per chain
and then is shared by all the shard tests. I'm sure that it's possible
to further improve this by compiling the binary once for _all_ chains,
but that requires more extensive rewrites to both workflows and I want
to keep these changes minimal.

As a bonus, I fixed a bug in the sourcify tests. Previously, if a
contract's metadata didn't include a `'remappings'` entry, then we would
fail to initialize the `ImportResolver` and the contract would be marked
"Unresolved". This fix initializes `ImportResolver::import_remaps` to an
empty `Vec` instead of failing, i.e. we assume that if `'remappings'` is
not present then there are no remappings. This appears to have fixed the
issue without causing any new failures.

---------

Co-authored-by: github-actions <github-actions@users.noreply.github.com>
Co-authored-by: Omar Tawfik <15987992+OmarTawfik@users.noreply.github.com>
Co-authored-by: Gustavo Giráldez <gustavo.giraldez@gmail.com>