Make the IR builder tolerant to parse errors (#1359)
If there are parse errors, the CST may contain nodes that don't respect
the structure in the language definition, and/or nodes that are of the
`UNRECOGNIZED` or `MISSING` kind. The IR builder that constructs the
structured AST from the CST needs to handle these cases gracefully if
we're to use its output to implement the new binder.

This changes the API of the builder functions to return an `Option<T>`
instead of a `Result<T>`. The builder code will now ignore some
malformed CST nodes and don't construct the corresponding AST node
(returning `None`), and will skip over `UNRECOGNIZED` or `MISSING`
nodes. It will panic on other cases of malformed CSTs. In general, the
code assumes the CST was generated by the parser for the language
definition.

Caveat: the new behaviour may hide builder logic errors.

---------

Co-authored-by: Omar Tawfik <15987992+OmarTawfik@users.noreply.github.com>