Fix #1375: untyped tuple deconstruction statements are assignments (#1377)
Fixes #1375 

Except for Solidity < 0.5.0 with the `var` keyword. If the keyword is
present, the statement should be treated as a declaration.

Caveat: because we don't validate code, it's possible to mix typed and
untyped tuple elements and the binder will create declarations and
references for each, but that's invalid Solidity code.