Initial new binder implementation (#1376)
This PR adds a series of backend passes to find all definitions and
references in a given `CompilationUnit`, as well as type all relevant
nodes. The information is output in a `Binder` object and a
`TypeRegistry`, which will record all reified Solidity types in the
unit.

Users can then query the `Binder` for definitions, references and types
of nodes in the CST/AST.

The implementation is still incomplete, but most of they foundations to
fully resolve and type the compilation unit are in place. The passes are
as follows:

- `p2_collect_definitions` (complete): collects all definitions in the
compilation unit
- `p3_type_definitions` (mostly complete): gather types and resolve all
types related to the definitions collected in the previous pass; this
pass also will collect some type references which are needed to fully
type the definitions (eg. the types of state variables, or the inherited
bases of a contract)
- `p4_resolve_references` (incomplete): collect and resolve the
remaining references and type all expression AST nodes

It is expected that we'll need a new short pass before p4 to linearise
the contracts and store that information in the definitions to use when
resolving symbols in the next pass.

We're also generating snapshot outputs for the new binder from the same
samples used for the graph implementation, but for now they are stored
in the `target` directory of the `snapshots` crate and not added to the
repository.