Customize recursion into contract definitions when resolving references
This is needed because the contract specifiers (inheritance and storage layout)
need to resolve in the parent scope, but the members should resolve in the
contract's scope.

An alternative solution to this is holding the contract scope ID and pushing it
delayed when entering the `ContractMembers` node. But this requires some
internal bookkepping and it seems more hackish.