Resolve members of undetermined typing by choosing the first type
This is because we cannot do argument-type lookup in this case (there are no
arguments). There should be a disambiguation in the resolution that leads to
this typing though. Eg. `this.foo.selector` with `foo()` being an overriden
method should choose the override definition. In fact in this case, the
resolution of `this.foo` should discard previous definitions from base classes.
There may be other overloads, but getting the selector in that case is an error.