Continue implementation of Solidity bindings rules in the new binder (#1405)
This PR implements most bindings rules in order to bring feature parity
with the existing stack graph binder, at least as far as the internal
bindings snapshots concern.

This mainly entails completing the implementation of the definition
typing and reference resolution passes. The final current pipeline
consists of 4 passes:

- `p2_collect_definitions` in which all definitions are collected with
their naming identifiers. Also lexical (and other kinds of) scopes are
identified and linked together, and definitions are registered into them
for later lookup.
- `p3_linearise_contracts` which will collect all bases of contracts and
interfaces and then compute the linearisation for each of them.
- `p4_type_definitions` which will determine and assign typing
information to all the definitions collected in `p2` and as a result
register most types used in the compilation unit. Additionally, public
state variables will be assigned an equivalent getter function type,
which happens after the main typing pass to ensure all types are already
registered.
- `p5_resolve_references` will find identifiers used as references,
resolve them to the appropriate definitions, and compute typing
information for AST nodes containing expressions and statements. Both
these actions are co-dependant and happen concurrently for each node.

All the information collected and computed by these passes are stored in
a `Binder` and `TypeRegistry` instances.

The `Binder` will store:
- `Definition` objects (identifier + definiens node), indexed by
definiens `NodeId`
- `Reference` objects (identifier + `Resolution`), indexed by identifier
`NodeId`
- `Scope` objects (set of definitions and relationship to other scopes),
indexed by `NodeId`
- index of root `FileScope`s for all files in the compilation unit
- `UsingDirective` objects registered both at valid scopes
(contracts/libraries/etc) and globally
- linearisations, as a vector of definitions, indexed by the
contract/interface definition's `NodeId`
- `Typing` information for each relevant `NodeId`

The `TypeRegistry` stores an index of registered types, both elementary
types and user defined types. Each type is given a `TypeId` for
efficient storage, lookup and comparison. The registry also stores some
common elementary types required to type most built-ins functions and
some kinds of expressions (eg. the boolean type).

There are two important value types that are central to how binding
works, and are closely related: `Resolution` and `Typing`.

`Resolution` represents the result of a lookup of a reference in the
context where that reference occurs. This can be:
- `Unresolved`, the identifier was not found.
- `Definition` to a single definition identified by a `NodeId`.
- `Ambiguous` when the lookup returns multiple possible definitions (eg.
multiple function overloads, or virtual methods in a contract
hierarchy). It's usually not possible to determine *which* of the found
definitions should apply, as more information is needed. Eg. the types
of arguments when the reference is used in a function call, to select
the appropriate overload.
- `BuiltIn` which means the symbol refers to a Solidity built-in of some
kind. The possible variants are encoded in an `enum` and the behaviour
of each is encoded in the `built_ins.rs` module.

`Typing` represents typing information related to an AST node. If the
node happens to be a reference identifier, there is a direct
relationship between its `Resolution` and the `Typing`. The possible
variants are:
- `Unresolved` the type of the node cannot be resolved at this time
(this normally propagates upwards).
- `Resolved` with a corresponding `TypeId` valid in the `TypeRegistry`.
- `Undetermined` when there are multiple possible types that result from
the typing of an `Ambiguous` resolution. Function call overload
resolution will select the appropriate type by matching the types of the
arguments and work backwards to the reference (if any) and fixup the
selected definition.
- `UserMetaType` is a node which refers to a user defined type by name,
eg. the identifier of a contract. This is not a type that can be
translated to the EVM, and as such is a way to suspend the typing until
more information on how it's used is gathered. Eg. to create a new
contract via the `new` operator, or construct a struct by using the
struct's name in a function call.
- `MetaType` similar to `UserMetaType`, but refers to the meta type of
an elementary type. A typical use case is explicit casting, which parses
as a function call.
- `BuiltIn` is used to type the `BuiltIn` resolution when the result is
not yet an actual type representable in the EVM. Eg. the built-in
function `addmod` resolves to `Resolution::BuiltIn(Addmod)` and will
type to `Typing::BuiltIn(Addmod)` which in turn will type to the
`uint256` type when evaluated in the context of a function call.
- `New` is a special typing that's used as the result of applying the
`new` operator, and will type to a contract reference when evaluated as
a function call.
- `This` and `Super` both refer to the typing of the `this` and `super`
keywords and they both alter the resolution rules when used in a member
access expression, or type to an `address` type equivalent when used as
a value.

The PR also implements some refactors as suggested in the code review of
the previous new binder PR #1376:
- re-exporting `NodeId` from instead of defining a new type
- changed import resolver API to work on `NodeId`s instead of `Cursor`s
- refactoring of `binder.rs` module and separating it into several
sub-modules
- built-ins related functionality now lives in its own module

---------

Co-authored-by: Beta Ziliani <beta@manas.tech>