More robust and better error reporting of errors in fetch (#1438)
Potentially fixes #1437 although the cause of the failure is unknown.
This patch should at least give us better insight to what is the cause.

EDIT: with several successful runs, it seems that the fix was being
stricter on the status code. Nevertheless, the rest of the code helps
better tracing potential issues that might arise.