Move runtime cargo (#1422)
Builds on top of #1420 and in turn of #1411 

It finishes moving `crates/codegen/runtime/cargo/crate` to the
`solidity_crate`. Notes on changes:

* The former `src/generated` folder that contained the cst, parser, etc
is renamed to `src/slang`.
* The generation of IR languages posed a challenge: their generation
can't be inlined, as from the same set of templates we generate various
IR languages. This PR replaces the logic to keep the same naming
convention that with the inlined templates: instead of
`<ir>/generated/<template_name>.rs` it generates
`<ir>/<template_name>.generated.rs`.
* There are no more stubs, therefore, a check is added in each IR
template to ensure that it's being generated with the proper model.
Otherwise, an empty string is generated, which isn't written into a
file.

~Note: It can be easily adapted to be independent of #1420 (but not of
#1411). It can also be split in two, with the last commit being in a
different PR. But It's probably OK as is.~