Overload (and other ambiguities) resolution and many fixes in the new binder (#1426)
This PR adds functionality to resolve function overloads and many other
ambiguities that were not possible to resolve with the stack graph
binder. It also fixes a lot of issues found via running the new binder
against contracts from sourcify (see #1419) as well as comparing the
bindings output by `solc`. The resulting binding information should be
much better than that obtained by the stack graph binder. Notably:

- function overloads are resolved by matching the type (and count) of
arguments
- `super` calls should resolve considering the current contract as the
target
- `using` directives consider the full type to which they apply
(previously eg. function types were not discernible)
- full and correct public getters types are created for complex data
types
- all built-ins should type correctly, even in tricky cases such as
`.push()` for arrays
- `abi.decode` can now be typed correctly

We're still not doing full validation, so the binder may return
incorrect results eg. for cases with no ambiguity but a type mismatch.
But the valid Solidity code, the expectation is that the new binder is
correct.

A lot of new snapshot cases are added in this PR, mostly to test
specific functionality or cases that the new binder handles. In many of
these, the old binder will produce incorrect output, so the diff reports
are expected to grow in number significantly.

---------

Co-authored-by: Beta Ziliani <beta@manas.tech>