Add clippy rule to enforce module style to use `mod.rs` (#1449)
Spin off from [this code review
thread](https://github.com/NomicFoundation/slang/pull/1419#discussion_r2414945437).

Since we're using Rust 1.87 and the rule implementation has bugs, we
need to override the rule in the `testing/sourcify` crate to avoid a
false positive. We also disable it in the `graph_builder` crate which is
an internal fork.