Refactor IR language modules generation (#1451)
Instead of running a separate process to render each IR language, make
each language invoke a render template for the modules it needs. This
should improve the efficiency of the `codegen` step and allows each
language to generate only the modules it requires.

This is a spin off from [this
comment](https://github.com/NomicFoundation/slang/pull/1422#discussion_r2372684321)
from a code review.