Simplify IR2 language (#1443)
This PR introduces several changes to the shape of the IR2 language used
to compute bindings and typing, to reduce the tree depth and improve its
usability.

- Function types are unified under a single IR2 type
`FunctionDefinition`, which now includes a `kind` discriminator. Regular
functions, constructors, modifiers, and special functions are now all
represented under this type in the IR2 tree.
- Collapsed several non-terminal types which consist of a single content
field and possibly punctuation or fixed unique keywords: parameters and
returns declarations, import aliases, else branches, index access
expressions, etc.
- Simplified arguments for function calls to a choice type with a vector
of expressions for positional parameters, or `NamedArgument` which is an
identifier and the corresponding expression.
- Simplified string literals, which no longer distinguish between single
and double quoted values. This applies to normal strings, hex strings
and unicode strings. Also, the `StringExpression` (for a string literal
value) is now a choice between vectors of strings, hex strings or
unicode strings and the (deprecated in newer Solidity) single value
variants are represented using one element vectors.
- Function and state variable attributes are flattened and categorized
into separate fields: visibility, mutability, override specifiers and
modifier invocations (for functions).
- All optional unique terminals are now represented as a boolean to
indicate the presence of the keyword.

As a result, the number of IR2 node types is reduced and the tree
structure is more shallow, making the passes code simpler.