[v2] add slang v2 lexer (#1462)
This final PR adds the actual lexer using the `logos` crate, and adds
some unit tests for it. Note that (partial) integration with sourcify
was added under `temp_sourcify.rs`, to be removed once the lexer is used
(and tested) through the parser instead.

There are some remaining work items to be done (grep for `TODO(v2)`
comments), but it can follow in later PRs.

> NOTE: this is a set of PRs to add the lexer v2, to make reviewing
smaller chunks easier:
>
> 1. https://github.com/NomicFoundation/slang/pull/1457
> 2. https://github.com/NomicFoundation/slang/pull/1458
> 3. https://github.com/NomicFoundation/slang/pull/1459
> 4. https://github.com/NomicFoundation/slang/pull/1460
> 5. https://github.com/NomicFoundation/slang/pull/1461
> 6. https://github.com/NomicFoundation/slang/pull/1462