Add benchmark test cases for the new binder (#1455)
This PR adds a benchmark case to measure the performance of the new
binder. I'd like to have a good baseline before making any big changes
in the implementation (eg. using red-green trees for the CST and IR
nodes). It would also be nice to have this merged before #1443 in order
to measure how the simplification of the IR used by the binder affects
performance.