Introduce `SemanticAnalysis` to gather the IR, binding and typing (#1476)
The new `SemanticAnalysis` object is the composition of the `Binder`,
`TypeRegistry` and the IR trees for all the files in the
`CompilationUnit`. Replaces the ad-hoc `Output` structure from the last
pass in the backend pipeline. It's created from the `CompilationUnit` in
a similar way to how the current `BindingGraph`.

NOTE: this is only the introduction of the type to provide a foundation
to build on. The API of `SemanticAnalysis` is not designed yet.